Fix for IA-64
git-svn-id: svn+ssh://mercurium.pc.ac.upc.edu/var/lib/svn/nanox/trunk@3426 39d7fdcd-100c-0410-a620-8454655e6f6e

+++ src/arch/unix-os/os.cpp
@@ -5,14 +5,14 @@ using namespace nanos;
 
 void OS::getProgramArguments (ArgumentList &list)
 {
-	int *p;
+	long *p;
 	int i;
 
 	int argc;
 	char **argv;
 
 	// variables are before environment
-	p=(int *)environ;
+	p=(long *)environ;
 
 	// go backwards until we find argc
 	p--;
